+++ mimetic/rfc822/group.cxx
@@ -33,7 +33,7 @@ Group::Group(const string& text)
 
 static string::size_type find_not_in_quote(const string& s, const string::value_type& c)
 {
-    int len = (int)s.length();
+    int len = s.length();
     bool in_dquote = false;
     for(int i =0; i < len; ++i)
     {
